requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#

asgiref==3.9.1
    # via
    #   django
    #   django-allauth
boto3==1.39.12
    # via django-storages
botocore==1.39.12
    # via
    #   boto3
    #   s3transfer
certifi==2025.7.14
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
cryptography==45.0.5
    # via
    #   djangorestframework-simplejwt
    #   pyjwt
dj-rest-auth[with-social,with_social]==7.0.1
    # via -r requirements/requirements.in
django==5.2.4
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth[socialaccount]==65.10.0
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
django-environ==0.12.0
    # via -r requirements/requirements.in
django-storages[s3]==1.14.6
    # via -r requirements/requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]==5.3.1
    # via -r requirements/requirements.in
gunicorn==23.0.0
    # via -r requirements/requirements.in
idna==3.10
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
oauthlib==3.3.1
    # via django-allauth
packaging==25.0
    # via gunicorn
pillow==11.3.0
    # via -r requirements/requirements.in
psycopg[binary]==3.2.3
    # via -r requirements/requirements.in
psycopg-binary==3.2.3
    # via psycopg
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.10.1
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via botocore
requests==2.32.4
    # via django-allauth
s3transfer==0.13.1
    # via boto3
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.14.1
    # via psycopg
tzdata==2025.2
    # via
    #   django
    #   psycopg
urllib3==2.5.0
    # via
    #   botocore
    #   requests
